requirements.txt
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.1.0
idna==2.10
openpyxl==3.0.7
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.4
